Create JAVA SCRIPT API for web cam access for object detection using open CV
This JAVA SCRIPT API acts as bridge between the hardware of our local system (aka PC, laptop, tablet etc.) and the cloud based Google Colab. Since Colab works on Google cloud 
the usual code in open CV is not sufficient for image capture. We have to use the above API and run it prior to executing the usual open CV code.
                     Further we can make changes on top of that API to create changes and variations in the image capturing and also add some more features.